Revised basics of how different this version is
This documentation is super basic. I need to come back to this and make proper write-ups of what has been changed and how it works. But the example output should make a lot of it obvious once you use the script or if you have experience with the original version.